+++ packages/apollo-client/src/ApolloClient.ts
@@ -14,12 +14,14 @@ import {
 } from 'apollo-utilities';
 
 import { QueryManager } from './core/QueryManager';
-import { ApolloQueryResult } from './core/types';
+import { ApolloQueryResult, OperationVariables } from './core/types';
 import { ObservableQuery } from './core/ObservableQuery';
 
 import { Observable } from './util/Observable';
 
 import {
+  QueryBaseOptions,
+  QueryOptions,
   WatchQueryOptions,
   SubscriptionOptions,
   MutationOptions,
@@ -33,7 +35,7 @@ import { version } from './version';
 
 export interface DefaultOptions {
   watchQuery?: ModifiableWatchQueryOptions;
-  query?: ModifiableWatchQueryOptions;
+  query?: QueryBaseOptions;
   mutate?: MutationBaseOptions;
 }
 
@@ -201,11 +203,16 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * a description of store reactivity.
    *
    */
-  public watchQuery<T>(options: WatchQueryOptions): ObservableQuery<T> {
+  public watchQuery<T, TVariables = OperationVariables>(
+    options: WatchQueryOptions<TVariables>,
+  ): ObservableQuery<T> {
     this.initQueryManager();
 
     if (this.defaultOptions.watchQuery) {
-      options = { ...this.defaultOptions.watchQuery, ...options };
+      options = {
+        ...this.defaultOptions.watchQuery,
+        ...options,
+      } as WatchQueryOptions<TVariables>;
     }
 
     // XXX Overwriting options is probably not the best way to do this long term...
@@ -214,26 +221,30 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
       (options.fetchPolicy === 'network-only' ||
         options.fetchPolicy === 'cache-and-network')
     ) {
-      options = { ...options, fetchPolicy: 'cache-first' } as WatchQueryOptions;
+      options = { ...options, fetchPolicy: 'cache-first' };
     }
 
     return this.queryManager.watchQuery<T>(options);
   }
 
   /**
-   * This resolves a single query according to the options specified and returns a
-   * {@link Promise} which is either resolved with the resulting data or rejected
-   * with an error.
+   * This resolves a single query according to the options specified and
+   * returns a {@link Promise} which is either resolved with the resulting data
+   * or rejected with an error.
    *
-   * @param options An object of type {@link WatchQueryOptions} that allows us to describe
-   * how this query should be treated e.g. whether it is a polling query, whether it should hit the
+   * @param options An object of type {@link QueryOptions} that allows us to
+   * describe how this query should be treated e.g. whether it should hit the
    * server at all or just resolve from the cache, etc.
    */
-  public query<T>(options: WatchQueryOptions): Promise<ApolloQueryResult<T>> {
+  public query<T, TVariables = OperationVariables>(
+    options: QueryOptions<TVariables>,
+  ): Promise<ApolloQueryResult<T>> {
     this.initQueryManager();
 
     if (this.defaultOptions.query) {
-      options = { ...this.defaultOptions.query, ...options };
+      options = { ...this.defaultOptions.query, ...options } as QueryOptions<
+        TVariables
+      >;
     }
 
     if (options.fetchPolicy === 'cache-and-network') {
@@ -242,9 +253,10 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
       );
     }
 
-    // XXX Overwriting options is probably not the best way to do this long term...
+    // XXX Overwriting options is probably not the best way to do this long
+    // term...
     if (this.disableNetworkFetches && options.fetchPolicy === 'network-only') {
-      options = { ...options, fetchPolicy: 'cache-first' } as WatchQueryOptions;
+      options = { ...options, fetchPolicy: 'cache-first' };
     }
 
     return this.queryManager.query<T>(options);
@@ -257,11 +269,16 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    *
    * It takes options as an object with the following keys and values:
    */
-  public mutate<T>(options: MutationOptions<T>): Promise<FetchResult<T>> {
+  public mutate<T, TVariables = OperationVariables>(
+    options: MutationOptions<T, TVariables>,
+  ): Promise<FetchResult<T>> {
     this.initQueryManager();
 
     if (this.defaultOptions.mutate) {
-      options = { ...this.defaultOptions.mutate, ...options };
+      options = {
+        ...this.defaultOptions.mutate,
+        ...options,
+      } as MutationOptions<T, TVariables>;
     }
 
     return this.queryManager.mutate<T>(options);
@@ -271,7 +288,9 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * This subscribes to a graphql subscription according to the options specified and returns an
    * {@link Observable} which either emits received data or an error.
    */
-  public subscribe<T = any>(options: SubscriptionOptions): Observable<T> {
+  public subscribe<T = any, TVariables = OperationVariables>(
+    options: SubscriptionOptions<TVariables>,
+  ): Observable<T> {
     this.initQueryManager();
 
     return this.queryManager.startGraphQLSubscription(options);
@@ -283,7 +302,7 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * the root query. To start at a specific id returned by `dataIdFromObject`
    * use `readFragment`.
    */
-  public readQuery<T, TVariables = any>(
+  public readQuery<T, TVariables = OperationVariables>(
     options: DataProxy.Query<TVariables>,
   ): T | null {
     return this.initProxy().readQuery<T>(options);
@@ -300,7 +319,7 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * in a document with multiple fragments then you must also specify a
    * `fragmentName`.
    */
-  public readFragment<T, TVariables = any>(
+  public readFragment<T, TVariables = OperationVariables>(
     options: DataProxy.Fragment<TVariables>,
   ): T | null {
     return this.initProxy().readFragment<T>(options);
@@ -311,7 +330,7 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * the store. This method will start at the root query. To start at a a
    * specific id returned by `dataIdFromObject` then use `writeFragment`.
    */
-  public writeQuery<TData = any, TVariables = any>(
+  public writeQuery<TData = any, TVariables = OperationVariables>(
     options: DataProxy.WriteQueryOptions<TData, TVariables>,
   ): void {
     const result = this.initProxy().writeQuery(options);
@@ -330,7 +349,7 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    * in a document with multiple fragments then you must also specify a
    * `fragmentName`.
    */
-  public writeFragment<TData = any, TVariables = any>(
+  public writeFragment<TData = any, TVariables = OperationVariables>(
     options: DataProxy.WriteFragmentOptions<TData, TVariables>,
   ): void {
     const result = this.initProxy().writeFragment(options);