+++ packages/graphql-anywhere/CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+### vNext
+
+- Add support for arrays to `filter`.       
+  [PR #3591](https://github.com/apollographql/apollo-client/pull/3591)
+
+### 4.1.14
+
+- No changes.
+
+### 4.1.13
+
+- No changes.
+
+### 4.1.12
+
+- No new changes.
+
 ### 4.1.11
 
 - Source files are now excluded when publishing to npm.